+++ Makefile
@@ -118,8 +118,9 @@ obj/bflib_mshandler.o \
 obj/bflib_mspointer.o \
 obj/bflib_netsession.o \
 obj/bflib_netsp.o \
-obj/bflib_network.o \
 obj/bflib_network_exchange.o \
+obj/net_main.o \
+obj/net_lobby.o \
 obj/net_resync.o \
 obj/bflib_planar.o \
 obj/bflib_render.o \
