+++ src/bflib_enet.cpp
@@ -15,8 +15,7 @@
 #include "pre_inc.h"
 #include "bflib_enet.h"
 #include "bflib_datetm.h"
-#include "bflib_network.h"
-#include "bflib_network_internal.h"
+#include "net_main.h"
 #include "front_network.h"
 #include "bflib_math.h"
 #include "net_portforward.h"
