+++ src/front_landview_multiplayer.c
@@ -23,7 +23,8 @@
 #include "bflib_basics.h"
 #include "bflib_datetm.h"
 #include "bflib_mouse.h"
-#include "bflib_network.h"
+#include "net_main.h"
+#include "net_lobby.h"
 #include "bflib_network_exchange.h"
 #include "bflib_sndlib.h"
 #include "bflib_sound.h"
