+++ src/net_game.c
@@ -23,7 +23,6 @@
 #include "globals.h"
 #include "bflib_basics.h"
 #include "bflib_coroutine.h"
-#include "bflib_network.h"
 #include "bflib_network_exchange.h"
 #include "net_resync.h"
