+++ src/net_input_lag.c
@@ -6,8 +6,7 @@
 #include "player_data.h"
 #include "net_game.h"
 #include "game_legacy.h"
-#include "bflib_network.h"
-#include "bflib_network_internal.h"
+#include "net_main.h"
 #include "bflib_enet.h"
 #include "bflib_datetm.h"
 #include "frontend.h"
