+++ src/net_lobby.c
@@ -1,0 +1,349 @@
+/******************************************************************************/
+// Free implementation of Bullfrog's Dungeon Keeper strategy game.
+/******************************************************************************/
+/** @file net_lobby.c
+ *     Lobby and session network support for Dungeon Keeper multiplayer.
+ */
+/******************************************************************************/
+#include "pre_inc.h"
+#include "net_lobby.h"
+
+#include "bflib_datetm.h"
+#include "bflib_enet.h"
+#include "bflib_network_exchange.h"
+#include "bflib_sound.h"
+#include "front_landview.h"
+#include "front_network.h"
+#include "net_lan.h"
+#include "net_matchmaking.h"
+#include "post_inc.h"
+
+#include <SDL2/SDL.h>
+/******************************************************************************/
+
+#define SESSION_COUNT 32
+
+static struct TbNetworkSessionNameEntry sessions[SESSION_COUNT];
+static int32_t server_port = 0;
+
+struct MatchmakingCreateTask {
+    uint16_t ipv4_port;
+    uint16_t ipv6_port;
+    char host_name[32];
+};
+
+static int matchmaking_create_thread(void *userdata)
+{
+    struct MatchmakingCreateTask *task = (struct MatchmakingCreateTask *)userdata;
+    if (matchmaking_connect() == 0) {
+        matchmaking_create(task->host_name, (int)task->ipv4_port, (int)task->ipv6_port);
+    }
+    free(task);
+    return 0;
+}
+
+static void AddSessionSegment(const char *start, const char *end)
+{
+    if (start == end) {
+        return;
+    }
+    for (unsigned i = 0; i < SESSION_COUNT; i += 1) {
+        if (sessions[i].in_use) {
+            continue;
+        }
+        sessions[i].in_use = 1;
+        sessions[i].joinable = 1;
+        size_t seglen = (size_t)(end - start);
+        net_copy_name_string(sessions[i].text, start, min((size_t)SESSION_NAME_MAX_LEN, seglen + 1));
+        return;
+    }
+}
+
+void LbNetwork_SendChatMessageImmediate(int player_id, const char *message)
+{
+    char *write_pos = begin_net_message(NETMSG_CHATMESSAGE);
+    *write_pos = player_id;
+    write_pos += 1;
+    strcpy(write_pos, message);
+    write_pos += strlen(message) + 1;
+    send_remote_buffer(write_pos);
+}
+
+void LbNetwork_SetServerPort(int port)
+{
+    server_port = port;
+}
+
+TbError process_login_message(NetUserId source, char *read_pos)
+{
+    if (source == SERVER_ID) {
+        netstate.my_id = (NetUserId)*read_pos;
+        read_pos += 1;
+        netstate.users[netstate.my_id].version = net_current_version;
+        netstate.users[SERVER_ID].version = *(const struct GameVersionPacket *)read_pos;
+        return Lb_OK;
+    }
+    if (netstate.users[source].progress != USER_CONNECTED) {
+        NETMSG("Peer was not in connected state");
+        return Lb_OK;
+    }
+    size_t max_read = sizeof(netstate.msg_buffer) - (read_pos - netstate.msg_buffer);
+    size_t password_len = strnlen(read_pos, max_read);
+    if (password_len >= max_read || password_len > sizeof(netstate.password)) {
+        NETDBG(6, "Connected peer sent invalid password");
+        netstate.sp->drop_user(source);
+        return Lb_OK;
+    }
+    if (netstate.password[0] != 0 && strncmp(read_pos, netstate.password, sizeof(netstate.password)) != 0) {
+        NETMSG("Peer chose wrong password");
+        return Lb_OK;
+    }
+    read_pos += password_len + 1;
+    max_read = sizeof(netstate.msg_buffer) - (read_pos - netstate.msg_buffer);
+    size_t name_len = strnlen(read_pos, max_read);
+    if (name_len == 0 || name_len >= max_read || name_len >= sizeof(netstate.users[source].name)) {
+        NETDBG(6, "Connected peer sent invalid name");
+        netstate.sp->drop_user(source);
+        return Lb_OK;
+    }
+    snprintf(netstate.users[source].name, sizeof(netstate.users[source].name), "%s", read_pos);
+    if (!isalnum(netstate.users[source].name[0])) {
+        NETDBG(6, "Connected peer had bad name starting with %c", netstate.users[source].name[0]);
+        netstate.sp->drop_user(source);
+        return Lb_OK;
+    }
+    read_pos += name_len + 1;
+    netstate.users[source].version = *(const struct GameVersionPacket *)read_pos;
+    NETMSG("User %s successfully logged in", netstate.users[source].name);
+    netstate.users[source].progress = USER_LOGGEDIN;
+    play_non_3d_sample(76);
+    char *write_pos = begin_net_message(NETMSG_LOGIN);
+    *write_pos = source;
+    write_pos += 1;
+    memcpy(write_pos, &netstate.users[SERVER_ID].version, sizeof(netstate.users[SERVER_ID].version));
+    write_pos += sizeof(netstate.users[SERVER_ID].version);
+    send_message_buffer(source, write_pos);
+    for (NetUserId user_id = 0; user_id < (NetUserId)netstate.max_players; user_id += 1) {
+        if (netstate.users[user_id].progress == USER_UNUSED) {
+            continue;
+        }
+        SendUserUpdate(source, user_id);
+        if (user_id != netstate.my_id && user_id != source) {
+            SendUserUpdate(user_id, source);
+        }
+    }
+    UpdateLocalPlayerInfo(source);
+    return Lb_OK;
+}
+
+TbError process_user_update_message(NetUserId source, char *read_pos)
+{
+    if (source != SERVER_ID) {
+        WARNLOG("Unexpected USERUPDATE");
+        return Lb_OK;
+    }
+    NetUserId user_id = (NetUserId)*read_pos;
+    if (user_id < 0 || user_id >= (NetUserId)netstate.max_players) {
+        ERRORLOG("Critical error: Out of range user ID %i received from server, could be used for buffer overflow attack", user_id);
+        abort();
+    }
+    read_pos += 1;
+    netstate.users[user_id].progress = (enum NetUserProgress)*read_pos;
+    read_pos += 1;
+    if (strnlen(read_pos, sizeof(netstate.msg_buffer) - (read_pos - netstate.msg_buffer)) >= sizeof(netstate.msg_buffer) - (read_pos - netstate.msg_buffer)) {
+        ERRORLOG("Critical error: Unterminated name in USERUPDATE");
+        abort();
+    }
+    snprintf(netstate.users[user_id].name, sizeof(netstate.users[user_id].name), "%s", read_pos);
+    UpdateLocalPlayerInfo(user_id);
+    return Lb_OK;
+}
+
+void LbNetwork_InitSessionsFromCmdLine(const char *str)
+{
+    NETMSG("Initializing sessions from command line: %s", str);
+    const char *start = str;
+    const char *end = str;
+    while (*end != '\0') {
+        if (start != end && (*end == ',' || *end == ';')) {
+            AddSessionSegment(start, end);
+            start = end + 1;
+        }
+        end += 1;
+    }
+    AddSessionSegment(start, end);
+}
+
+TbError LbNetwork_ExchangeLogin(char *plyr_name)
+{
+    NETMSG("Logging in as %s", plyr_name);
+    if (1 + strlen(netstate.password) + 1 + strlen(plyr_name) + 1 + sizeof(net_current_version) >= sizeof(netstate.msg_buffer)) {
+        ERRORLOG("Login credentials too long");
+        return Lb_FAIL;
+    }
+    char *write_pos = begin_net_message(NETMSG_LOGIN);
+    strcpy(write_pos, netstate.password);
+    write_pos += strlen(netstate.password) + 1;
+    strcpy(write_pos, plyr_name);
+    write_pos += strlen(plyr_name) + 1;
+    memcpy(write_pos, &net_current_version, sizeof(net_current_version));
+    write_pos += sizeof(net_current_version);
+    send_message_buffer(SERVER_ID, write_pos);
+    TbClockMSec start = LbTimerClock();
+    while (true) {
+        TbClockMSec elapsed = LbTimerClock() - start;
+        if (elapsed >= TIMEOUT_JOIN_LOBBY) {
+            NETMSG("ExchangeLogin: timed out waiting for login response (%dms)", (int)TIMEOUT_JOIN_LOBBY);
+            break;
+        }
+        unsigned wait_ms = (unsigned)(TIMEOUT_JOIN_LOBBY - elapsed);
+        if (!netstate.sp->msgready(SERVER_ID, wait_ms)) {
+            NETMSG("ExchangeLogin: msgready returned false");
+            break;
+        }
+        if (ProcessMessage(SERVER_ID, &net_screen_packet, sizeof(struct ScreenPacket)) == Lb_FAIL) {
+            NETMSG("ExchangeLogin: ProcessMessage failed");
+            break;
+        }
+        if (netstate.msg_buffer[0] == NETMSG_LOGIN) {
+            break;
+        }
+    }
+    if (netstate.msg_buffer[0] != NETMSG_LOGIN) {
+        NETMSG("ExchangeLogin: login rejected (msg_buffer[0]=%d)", (int)netstate.msg_buffer[0]);
+        return Lb_FAIL;
+    }
+    if (netstate.sp->msgready(SERVER_ID, TIMEOUT_JOIN_LOBBY)) {
+        ProcessMessage(SERVER_ID, &net_screen_packet, sizeof(struct ScreenPacket));
+    }
+    if (netstate.my_id == INVALID_USER_ID) {
+        NETMSG("ExchangeLogin: login unsuccessful, still INVALID_USER_ID");
+        return Lb_FAIL;
+    }
+    return Lb_OK;
+}
+
+TbError LbNetwork_Create(char *, char *plyr_name, uint32_t *plyr_num, void *optns)
+{
+    if (!netstate.sp) {
+        ERRORLOG("No network SP selected");
+        return Lb_FAIL;
+    }
+    char default_port_buf[16];
+    snprintf(default_port_buf, sizeof(default_port_buf), ":%u", (unsigned)ENET_DEFAULT_PORT);
+    const char *port = default_port_buf;
+    char port_string[16] = "";
+    if (server_port != 0) {
+        snprintf(port_string, sizeof(port_string), "%d", server_port);
+        port = port_string;
+    }
+    if (netstate.sp->host(port, optns) == Lb_FAIL) {
+        return Lb_FAIL;
+    }
+    uint16_t local_port = ENET_DEFAULT_PORT;
+    if (server_port > 0) {
+        local_port = (uint16_t)server_port;
+    }
+    uint16_t ipv4_port = local_port;
+    if (external_ipv4_port != 0) {
+        ipv4_port = external_ipv4_port;
+    }
+    const uint16_t ipv6_port = enet_get_bound_ipv6_port();
+    if (frontnet_service_selected(FrontendNetSvc_LAN)) {
+        lan_host_start(plyr_name, local_port);
+    }
+    if (frontnet_service_selected(FrontendNetSvc_Online)) {
+        struct MatchmakingCreateTask *task = malloc(sizeof(struct MatchmakingCreateTask));
+        if (task != NULL) {
+            task->ipv4_port = ipv4_port;
+            task->ipv6_port = ipv6_port;
+            snprintf(task->host_name, sizeof(task->host_name), "%s", plyr_name);
+            SDL_Thread *thread = SDL_CreateThread(matchmaking_create_thread, "matchmaking_host", task);
+            if (thread != NULL) {
+                SDL_DetachThread(thread);
+            } else {
+                free(task);
+            }
+        }
+    }
+    netstate.my_id = SERVER_ID;
+    snprintf(netstate.users[netstate.my_id].name, sizeof(netstate.users[netstate.my_id].name), "%s", plyr_name);
+    netstate.users[netstate.my_id].progress = USER_LOGGEDIN;
+    netstate.users[netstate.my_id].version = net_current_version;
+    *plyr_num = netstate.my_id;
+    UpdateLocalPlayerInfo(netstate.my_id);
+    LbNetwork_EnableNewPlayers(true);
+    return Lb_OK;
+}
+
+TbError LbNetwork_Join(struct TbNetworkSessionNameEntry *nsname, char *plyr_name, int32_t *plyr_num, void *optns)
+{
+    if (!netstate.sp) {
+        ERRORLOG("No network SP selected");
+        return Lb_FAIL;
+    }
+    if (netstate.sp->join(nsname->text, optns) == Lb_FAIL) {
+        return Lb_FAIL;
+    }
+    netstate.my_id = INVALID_USER_ID;
+    if (LbNetwork_ExchangeLogin(plyr_name) == Lb_FAIL) {
+        return Lb_FAIL;
+    }
+    *plyr_num = netstate.my_id;
+    return Lb_OK;
+}
+
+TbError LbNetwork_EnableNewPlayers(TbBool allow)
+{
+    if (!netstate.locked && !allow) {
+        for (NetUserId id = 0; id < (NetUserId)netstate.max_players; id += 1) {
+            if (netstate.users[id].progress == USER_CONNECTED) {
+                netstate.sp->drop_user(id);
+            }
+        }
+    }
+    netstate.locked = !allow;
+    if (netstate.locked) {
+        NETMSG("New players are NOT allowed to join");
+    } else {
+        NETMSG("New players ARE allowed to join");
+    }
+    return Lb_OK;
+}
+
+TbError LbNetwork_Stop(void)
+{
+    lan_shutdown();
+    matchmaking_disconnect();
+    if (netstate.sp) {
+        netstate.sp->exit();
+    }
+    memset(&netstate, 0, sizeof(netstate));
+    netstate.my_id = INVALID_USER_ID;
+    return Lb_OK;
+}
+
+TbError LbNetwork_EnumeratePlayers(struct TbNetworkSessionNameEntry *, TbNetworkCallbackFunc callback, void *buf)
+{
+    struct TbNetworkCallbackData data;
+    for (NetUserId id = 0; id < (NetUserId)netstate.max_players; id += 1) {
+        if (!IsUserActive(id)) {
+            continue;
+        }
+        memset(&data, 0, sizeof(data));
+        snprintf(data.plyr_name, sizeof(data.plyr_name), "%s", netstate.users[id].name);
+        callback(&data, buf);
+    }
+    return Lb_OK;
+}
+
+TbError LbNetwork_EnumerateSessions(TbNetworkCallbackFunc callback, void *ptr)
+{
+    for (unsigned i = 0; i < SESSION_COUNT; i += 1) {
+        if (!sessions[i].in_use) {
+            continue;
+        }
+        callback((struct TbNetworkCallbackData *)&sessions[i], ptr);
+    }
+    return Lb_OK;
+}
