+++ src/net_received_packets.c
@@ -21,7 +21,7 @@
 
 #include "globals.h"
 #include "bflib_basics.h"
-#include "bflib_network.h"
+#include "net_main.h"
 #include "packets.h"
 #include "player_data.h"
 #include "game_legacy.h"
